Initial implementation of Concourse resource wrapper
This project reduces the boilerplate required to implement a
Concourse resource. To use it, a `Resource` interface must be
implemented, which has three methods `Check`, `In`, and `Out`.
These will become the commands /opt/resource/check,
/opt/resource/in, and /opt/resource/out that Concourse calls
during the life cycle of the resource.

See https://concourse-ci.org/implementing-resources.html for full
documentation on implementing a Concourse resource.

With the interface's methods in place, this library will handle:

* parsing of JSON from standard input and passing it to the methods
  as Go values. Most of the parsed input is passed as maps to the
  methods.
* converting the outputs of the methods back to JSON, and printing
  the values to standard output as expected by Concourse. The
  methods return maps and arrays and do not need to concern
  themselves with marshaling to JSON.
* logging to standard error, to prevent corrupting the JSON output
  expected on stdout. Each method receives a logger argument with
  builtin log levels and color coding of messages for the Concourse
  UI.

The Resource methods may be tested easily, as they do not need to
deal with standard input and output or parsing of JSON, they
simply receive input as arguments, and they return values as output.