Use Metadata type instead of using a map
The metadata type that Concourse expects on output is an array of
fixed key value pairs with keys `Name` and `Value`. Because these
will not change, it is not necessary to use a map, which was being
used for its dynamic nature. Using a map also prevents outputting
duplicate names, which is possible with the fixed key value pairs,
and may be useful for some resources. A map will still be used to
represent versions, which may have arbitrary keys.