Add project skeleton generator
This adds a command called `ofcourse` which may be run to generate
a skeleton project for a new Concourse resource which uses this
library. After running this command, the user should have a
minimally working resource which can be extended to do something
useful. In addition to the sample code, the project includes a
Dockerfile to package the resource, a Makefile, and a sample
pipeline which uses the resource.