Add the environment as an argument to resource functions
This adds a new interface which represents the environment of a
process, and it is passed as one of the arguments to each of
Check, In, and Out functions. This enables easier testing by being
able to change the environment via arguments instead of having to
read from e.g. `os.Environ()` within the functions.